# backend-arludent/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Arludent - Sistema de Gestión Odontológica",
        "description": "API RESTful completa para gestión de consultorios odontológicos con autenticación JWT, MFA y módulos clínicos",
        "contact": {
            "name": "Soporte Técnico Arludent",
            "email": "soporte@arludent.com"
        },
        "license": {
            "name": "Propietario",
            "url": "https://arludent.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Servidor de Desarrollo Local"
        },
        {
            "url": "https://api.arludent.com",
            "description": "Servidor de Producción"
        }
    ],
    "paths": {
...
        "/api/auth/recuperar-password": {
            "post": {
                "tags": [
                    "Autenticación"
                ],
                "summary": "Solicitar recuperación de contraseña",
                "description": "Solicita recuperación de contraseña",
                "operationId": "2b04fcfb2d23fdf809c1ae241d1b6c09",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "correo"
                                ],
                                "properties": {
                                    "correo": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "juan@example.com"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Correo de recuperación enviado"
                    }
                }
            }
        },
        "/api/auth/restablecer-password": {
            "post": {
                "tags": [
                    "Autenticación"
                ],
                "summary": "Restablecer contraseña",
                "description": "Restablece la contraseña con el token",
                "operationId": "b2544e43fa9d57a2a93632c3f54ed20b",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "correo",
                                    "token",
                                    "password",
                                    "password_confirmation"
                                ],
                                "properties": {
                                    "correo": {
                                        "type": "string",
                                        "format": "email"
                                    },
                                    "token": {
                                        "type": "string"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password"
                                    },
                                    "password_confirmation": {
                                        "type": "string",
                                        "format": "password"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Contraseña restablecida exitosamente"
                    }
                }
            }
        },
        "/api/auth/cambiar-password": {
            "post": {
                "tags": [
                    "Autenticación"
                ],
                "summary": "Cambiar contraseña",
                "description": "Cambia la contraseña del usuario autenticado",
                "operationId": "ae7af611caab6d201fbeb803d43e7151",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "password_actual",
                                    "password",
                                    "password_confirmation"
                                ],
                                "properties": {
                                    "password_actual": {
                                        "type": "string",
                                        "format": "password"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password"
                                    },
                                    "password_confirmation": {
                                        "type": "string",
                                        "format": "password"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Contraseña cambiada exitosamente"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearer": {
                "type": "http",
                "description": "Autenticación mediante token JWT. Formato: Bearer {token}",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}